# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Filmoteka API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/actor": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "CreateActor",
                "operationId": "create-actor",
                "parameters": [
                    {
                        "description": "List Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/actors/movies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "GetAllActorsWithMovies",
                "operationId": "get-actor-movie",
                "responses": {}
            }
        },
        "/actors/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "UpdateActorById",
                "operationId": "update-actor",
                "parameters": [
                    {
                        "description": "Actor Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "actors"
                ],
                "summary": "DeleteActorById",
                "operationId": "delete-actor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "PartiallyUpdateActorById",
                "operationId": "partially-update-actor",
                "parameters": [
                    {
                        "description": "Actor Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Actor"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "tags": [
                    "movies"
                ],
                "summary": "SignInUser",
                "operationId": "sing-in-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "SignUpUser",
                "operationId": "sing-up-user",
                "parameters": [
                    {
                        "description": "User Auth",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserAuth"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/movie": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "CreateMovie",
                "operationId": "create-movie",
                "parameters": [
                    {
                        "description": "Movie Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "tags": [
                    "actors"
                ],
                "summary": "GetMoviesByName",
                "operationId": "get-sorted-movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sort by query parram",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/movies/actors/{actor_name}": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "GetMoviesByName",
                "operationId": "get-movies-by-actor-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Actor Name",
                        "name": "actor_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "UpdateMovieById",
                "operationId": "update-movie",
                "parameters": [
                    {
                        "description": "Movie Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "movies"
                ],
                "summary": "DeleteMovieById",
                "operationId": "delete-movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "PartiallyUpdateMovieById",
                "operationId": "partially-update-movie",
                "parameters": [
                    {
                        "description": "Actor Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Actor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "desctiption": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "models.UserAuth": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}